Formatting new script variable and new event
You can ignore the Formatting, that was VS Code not me :)
I've put the $files variable into the script scope so I can use it later.
I've put a new event $wpf.dgUser.add_SelectedCellsChanged
and used it to grab the content of the cell, match it to the contents of
$files and create the log file path
I then use the default OS program to open the path.